+++ employeeservice/src/main/java/com/rms/dao/EmployeeDao.java
@@ -1,0 +1,9 @@
+package com.rms.dao;
+
+import org.springframework.data.repository.CrudRepository;
+
+import com.rms.model.*;
+
+public interface EmployeeDao extends CrudRepository <Employee, Integer> {
+    
+}
