+++ employeeservice/src/main/java/com/rms/model/Department.java
@@ -1,0 +1,22 @@
+package com.rms.model;
+
+
+/**
+ * This is a department
+ */
+public enum Department {
+
+
+    /**
+     * stored as ints in RDBMS and converted to String type here
+     */
+    TRAINING(1), STAGING(2), QC(3), RETENTINON(4), HR(5);
+
+	private final int type;
+
+	private Department(int type){this.type = type;}
+
+	public int getValue() {
+		return type;
+	}
+}
