+++ employeeservice/src/main/java/com/rms/service/EmployeeService.java
@@ -1,10 +1,29 @@
 package com.rms.service;
 
+import java.util.List;
+
+import com.rms.dao.EmployeeDao;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
 import com.rms.model.Employee;
 
-import org.springframework.data.repository.CrudRepository;
+/**
+ * Employee Service
+ */
+@Service
+public class EmployeeService {
 
-public interface EmployeeService extends CrudRepository <Employee, Integer> {
+    @Autowired
+    EmployeeDao ed;
 
-    
+    public void save(Employee emp) {
+        ed.save(emp);
+    }
+
+    public void delete(Employee emp) {
+        ed.delete(emp);
+    }
+
 }
