+++ employeeservice/src/test/java/com/rms/employeeservice/DemoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.revychan.employeeservice;
+package com.rms.employeeservice;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
